+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 ## [1.16.0](https://github.com/supercharge/strings/compare/v1.15.1...v1.16.0) - 2020-10-xx
 
 ### Added
-- `notIncludes(needle)`: determine whether the string contains the given `needle`
+- `notEquals(value)`: determine whether the string does not equal the given `value`
+- `notIncludes(needle)`: determine whether the string does not contain the given `needle`
 - `includesAll(...needles)`: determine whether the string contains all of the given `needles`
 
 ### Updated